Rename router folder
